va_arg format

+++ ft_printf_x.c
@@ -6,11 +6,12 @@
 /*   By: lpinheir <lpinheir@student.42sp.org.b      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/26 17:25:38 by lpinheir          #+#    #+#             */
-/*   Updated: 2021/04/29 15:09:00 by lpinheir         ###   ########.fr       */
+/*   Updated: 2021/05/01 11:20:58 by lpinheir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
+#include <stdio.h>
 
 static int	print_x(int precision, int strlen, long int num, char *hex)
 {
@@ -56,7 +57,7 @@ int	ft_printf_x(int inverted, int width, int precision, va_list args)
 {
 	long int	num;
 
-	num = va_arg(args, long int);
+	num = va_arg(args, int);
 	if (width < 0)
 	{
 		if (precision >= 0)
